Convert FauxComponentView to use React
