+++ examples/hosts/hosts-sample/package.json
@@ -37,11 +37,13 @@
     "@fluidframework/test-runtime-utils": "^0.55.0",
     "@fluidframework/view-interfaces": "^0.55.0",
     "@fluidframework/web-code-loader": "^0.55.0",
+    "react": "^16.10.2",
     "semver": "^7.3.4"
   },
   "devDependencies": {
     "@fluidframework/build-common": "^0.23.0",
     "@fluidframework/eslint-config-fluid": "^0.24.0",
+    "@types/react": "^16.9.15",
     "@types/semver": "^7.3.6",
     "@typescript-eslint/eslint-plugin": "~4.14.0",
     "@typescript-eslint/parser": "~4.14.0",
